requirements.txt
# apt install python3-tk
pygame~=1.9.6
# https://forums.developer.nvidia.com/t/install-pygame-on-jetson-nano/83731/6
# sudo apt-get install python-dev libsdl-image1.2-dev libsdl-mixer1.2-dev libsdl-ttf2.0-dev libsdl1.2-dev libsmpeg-dev python-numpy subversion libportmidi-dev ffmpeg libswscale-dev libavformat-dev libavcodec-dev
# sudo apt-get install libfreetype6-dev
getch
pyserial~=3.4
numpy~=1.19.0
pigpio~=1.46
pillow